+++ backupkit/src/commonTest/kotlin/com/vocabloot/backupkit/SyncEngineTest.kt
@@ -1,5 +1,9 @@
 package com.vocabloot.backupkit
 
+import com.vocabloot.backupkit.test.FakeCloudStorage
+import com.vocabloot.backupkit.test.MemoryRestoreRecordStore
+import com.vocabloot.backupkit.test.MemorySyncStateStore
+
 import kotlinx.coroutines.test.runTest
 import kotlin.test.Test
 import kotlin.test.assertEquals